flask_app/requirements.txt
flask
pyyaml
matplotlib
rasterio
torch
